requirements.txt
ortools==7.8.7959
Pillow==7.2.0
